core: align stack on a 32bit boundary
Fixes #1267.

+++ core/include/thread.h
@@ -140,7 +140,7 @@ int thread_getpid(void);
  *
  * @return          the amount of unused space of the thread's stack
  */
-int thread_measure_stack_free(char *stack);
+uintptr_t thread_measure_stack_free(char *stack);
 
 /* @} */
 #endif /* __THREAD_H */
